proper library interface and cleanup
