+++ extraction/run_verus.sh
@@ -35,7 +35,7 @@ fi
 
 src_path=$ext_dir_path/src
 
-main_file=$src_path/main.rs
+main_file=$src_path/lib.rs
 
 if [ $format = "true" ]; then
     echo ""
@@ -49,15 +49,15 @@ pushd $VESTPATH
 make
 popd
 
-echo ""
-echo "COMPILING LIB FILE" 
-pushd $ext_dir_path 
-cargo build --lib
-popd
+# echo ""
+# echo "COMPILING LIB FILE" 
+# pushd $ext_dir_path 
+# cargo build --lib
+# popd
 
 echo ""
 echo "VERIFYING" 
-verus -L dependency=$(realpath $ext_dir_path/target/debug/deps) $( find $ext_dir_path/target/debug/deps -name \*.rlib -exec realpath '{}' ';' | awk -F/ '{print "--extern " substr ($NF,4,index($NF,"-") - 4) "=" $0}' | grep -v vstd | grep -v builtin ) --extern vest=$VESTPATH/libvest.rlib --import vest=$VESTPATH/vest.verusdata --multiple-errors=100 --rlimit=1000 $main_file $verus_args -V spinoff-all 
+verus -L dependency=$(realpath $ext_dir_path/target/debug/deps) $( find $ext_dir_path/target/debug/deps -name \*.rlib -exec realpath '{}' ';' | awk -F/ '{print "--extern " substr ($NF,4,index($NF,"-") - 4) "=" $0}' | grep -v vstd | grep -v builtin ) --extern vest=$VESTPATH/libvest.rlib --import vest=$VESTPATH/vest.verusdata --multiple-errors=100 --rlimit=1000 --crate-type=lib $main_file $verus_args -V spinoff-all 
 # verus -L dependency=$(realpath $ext_dir_path/target/debug/deps) $( find $ext_dir_path/target/debug/deps -name \*.rlib -exec realpath '{}' ';' | awk -F/ '{print "--extern " substr ($NF,4,index($NF,"-") - 4) "=" $0}' | grep -v vstd | grep -v builtin ) --import vest=$VESTPATH/vest.verusdata --multiple-errors=100 --rlimit=100 $main_file $verus_args -V spinoff-all
